# internal/model/settings.go
// Package model provides domain entities for the resource management system.
// input: time package
// output: Environment, Owner, Role structs
// pos: Reference data entities for organizational context
// note: if this file changes, update header and README.md
package model

import "time"

type Environment struct {
	ID          string    `json:"id"`
	Name        string    `json:"name"`
	Slug        string    `json:"slug"`
	Description string    `json:"description"`
	CreatedAt   time.Time `json:"createdAt"`
}

type Owner struct {
	ID        string    `json:"id"`
	Name      string    `json:"name"`
	Email     string    `json:"email"`
	CreatedAt time.Time `json:"createdAt"`
}

type Role struct {
	ID          string    `json:"id"`
	Name        string    `json:"name"`
	Description string    `json:"description"`
	CreatedAt   time.Time `json:"createdAt"`
}