# internal/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ControlHub API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /auth/login:
    post:
      summary: Login with local credentials
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
  /resources:
    get:
      summary: List resources
      operationId: listResources
      parameters:
        - in: query
          name: type
          schema:
            $ref: "#/components/schemas/ResourceType"
        - in: query
          name: environmentId
          schema:
            type: string
      responses:
        "200":
          description: Resource list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Resource"
  /resources/{id}:
    get:
      summary: Get resource detail
      operationId: getResource
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resource detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /resources/{id}/profile:
    get:
      summary: Get resource typed profile projection
      description: >
        Read-only projection of the typed profile row for a resource.
        The `profile` object keys vary by `resourceType` and
        `resourceSubtype`.  Values may be string, number, boolean, or null.
        If the resource exists but has no typed profile row, `profile` is
        returned as an empty object `{}`.
      operationId: getResourceProfile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resource profile projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceProfileResponse"
        "404":
          description: Resource not found
  /resources/{id}/relations:
    get:
      summary: List relations for a resource
      operationId: listResourceRelations
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resource relation list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ResourceRelation"
  /resources/{id}/audit-events:
    get:
      summary: List audit events for a resource
      description: >
        Phase-1 bootstrap endpoint. Backed by a minimal MySQL table for
        local development only; the long-term backing store will be
        ClickHouse.  The response shape will remain stable after migration.
      operationId: listResourceAuditEvents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resource audit list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditEvent"
  /audit-events:
    get:
      summary: List audit events
      description: >
        Phase-1 bootstrap endpoint. Backed by a minimal MySQL table for
        local development only; the long-term backing store will be
        ClickHouse.  The response shape will remain stable after migration.
      operationId: listAuditEvents
      responses:
        "200":
          description: Audit event list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditEvent"
  /environments:
    get:
      summary: List environments
      operationId: listEnvironments
      responses:
        "200":
          description: Environment list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Environment"
  /owners:
    get:
      summary: List owners
      operationId: listOwners
      responses:
        "200":
          description: Owner list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Owner"
  /roles:
    get:
      summary: List roles
      operationId: listRoles
      responses:
        "200":
          description: Role list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Role"
  /resource-types:
    get:
      summary: List resource types
      operationId: listResourceTypes
      responses:
        "200":
          description: Resource type dictionary
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/DictionaryItem"
  /relation-types:
    get:
      summary: List relation types
      operationId: listRelationTypes
      responses:
        "200":
          description: Relation type dictionary
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/DictionaryItem"
  /lifecycle-statuses:
    get:
      summary: List lifecycle statuses
      operationId: listLifecycleStatuses
      responses:
        "200":
          description: Lifecycle status dictionary
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/DictionaryItem"
  /health-statuses:
    get:
      summary: List health statuses
      operationId: listHealthStatuses
      responses:
        "200":
          description: Health status dictionary
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/DictionaryItem"
  # ──────────────────────────────────────────────────────────────
  # PROPOSED: Topology endpoint contract (not yet implemented)
  #
  # GET /resources/{id}/topology
  #
  # Returns a local topology projection around a resource.
  # Scope: 1-hop or 2-hop neighborhood, built from existing
  # resources + relations. Not a global graph.
  #
  # Proposed response shape:
  #   {
  #     "rootResourceId": "string",
  #     "nodes": [
  #       {
  #         "id": "string",
  #         "resourceType": "string",
  #         "name": "string",
  #         "displayName": "string",
  #         "lifecycleStatus": "string",
  #         "healthStatus": "string"
  #       }
  #     ],
  #     "edges": [
  #       {
  #         "id": "string",
  #         "fromResourceId": "string",
  #         "toResourceId": "string",
  #         "relationType": "string"
  #       }
  #     ],
  #     "groups": [
  #       {
  #         "id": "string",
  #         "label": "string",
  #         "nodeIds": ["string"]
  #       }
  #     ]
  #   }
  #
  # Design notes:
  # - nodes/edges are required; groups is optional
  # - groups provide layout hints (e.g., cluster membership)
  # - no layout engine; positioning is client-side
  # - no topology persistence; built on-the-fly from resources + relations
  # - rootResourceId identifies the center of the local graph
  # ──────────────────────────────────────────────────────────────
components:
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      required: [token, role]
      properties:
        token:
          type: string
        role:
          type: string
    ResourceType:
      type: string
      enum:
        - host
        - database_instance
        - database_cluster
        - service
        - domain_name
        - virtual_ip
        - database_proxy
        - control_plane_component
    RelationType:
      type: string
      enum:
        - depends_on
        - member_of
        - runs_on
        - points_to
        - fronts
        - manages
        - replicates_to
    LifecycleStatus:
      type: string
      enum:
        - provisioning
        - running
        - stopped
        - degraded
        - decommissioning
    HealthStatus:
      type: string
      enum:
        - healthy
        - warning
        - critical
        - unknown
    DictionaryItem:
      type: object
      required: [key, label, description]
      properties:
        key:
          type: string
        label:
          type: string
        description:
          type: string
    Resource:
      type: object
      required:
        - id
        - resourceType
        - resourceSubtype
        - name
        - displayName
        - environmentId
        - ownerId
        - lifecycleStatus
        - healthStatus
        - source
        - externalId
      properties:
        id:
          type: string
        resourceType:
          $ref: "#/components/schemas/ResourceType"
        resourceSubtype:
          type: string
        name:
          type: string
        displayName:
          type: string
        environmentId:
          type: string
        ownerId:
          type: string
        lifecycleStatus:
          $ref: "#/components/schemas/LifecycleStatus"
        healthStatus:
          $ref: "#/components/schemas/HealthStatus"
        source:
          type: string
        externalId:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ResourceRelation:
      type: object
      required: [id, fromResourceId, toResourceId, relationType, createdAt]
      properties:
        id:
          type: string
        fromResourceId:
          type: string
        toResourceId:
          type: string
        relationType:
          $ref: "#/components/schemas/RelationType"
        createdAt:
          type: string
          format: date-time
    AuditEvent:
      type: object
      required: [id, actorUserId, eventType, result, createdAt]
      properties:
        id:
          type: string
        actorUserId:
          type: string
        targetResourceId:
          type: string
        eventType:
          type: string
        result:
          type: string
        createdAt:
          type: string
          format: date-time
    Environment:
      type: object
      required: [id, name, slug, description, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    Owner:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
    Role:
      type: object
      required: [id, name, description, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    ResourceProfileResponse:
      type: object
      required: [resourceId, resourceType, resourceSubtype, profile]
      properties:
        resourceId:
          type: string
        resourceType:
          $ref: "#/components/schemas/ResourceType"
        resourceSubtype:
          type: string
        profile:
          type: object
          additionalProperties: true
          description: >
            Typed profile fields.  Keys depend on resourceType:
            - database_instance: engine, version, host, port, role
            - database_cluster: engine, topologyMode, primaryEndpoint
            - service: systemName, repositoryUrl, runtimeEnv
            - host: hostname, ipAddress, osName